Moved test files to debug and added back the commit_from_git.txt which is used to test with SCMCommitsTests.ts
